+++ providers/pagerduty/src/airflow/providers/pagerduty/hooks/pagerduty.py
@@ -24,11 +24,7 @@ from typing import Any
 import pagerduty
 
 from airflow.exceptions import AirflowException
-
-try:
-    from airflow.sdk import BaseHook
-except ImportError:
-    from airflow.hooks.base import BaseHook  # type: ignore[attr-defined,no-redef]
+from airflow.providers.pagerduty.version_compat import BaseHook
 
 
 class PagerdutyHook(BaseHook):
